back/requirements.txt
flask
requests
flask-cors
Flask-PyMongo
wheel
flask-mongoengine
pyjwt
python-dateutil
redis
xmltodict
feedparser
